Add options to osmdbt-get-log to ignore changes and treat an empty result as OK.
Added the `--ignore-earlier-than LSN` command line flag, which ignores changes up to and including `LSN`. The PostgreSQL documentation explicitly warns that database system crashes (and perhaps failover to a replica?) can cause the loss of the `catchup` commit message, which means PostgreSQL will repeat log messages we've already seen. We get to choose whether we want to deal with gaps (i.e: when we crash after committing the catchup, but before writing external files) or duplicates (i.e: when we crash before committing the catchup, but after writing external files). Doing the latter, plus storing the LSN in the external `state.txt`, allows us to be robust without missing changes or getting duplicates.

Added the `--empty-is-ok` command line flag, which changes the program behaviour exit with a code of zero when there are no changes to output from the database. Normal behaviour is to exit with an error code. An empty set of changes is a common result (if only in development) which needs special handling, although not the "panic-and-email" that most errors should. Being able to treat all non-zero exit codes as errors simplifies downstream code using `osmdbt-get-log`.